Update unit test coverage for connection workers
