+++ publishing-sdk/src/test/java/com/ably/tracking/publisher/workerqueue/workers/AddTrackableWorkerTest.kt
@@ -3,13 +3,18 @@ package com.ably.tracking.publisher.workerqueue.workers
 import com.ably.tracking.TrackableState
 import com.ably.tracking.common.Ably
 import com.ably.tracking.common.ResultCallbackFunction
+import com.ably.tracking.publisher.DefaultCorePublisher
+import com.ably.tracking.publisher.PublisherInteractor
 import com.ably.tracking.publisher.Trackable
 import com.ably.tracking.publisher.workerqueue.WorkerSpecification
 import com.ably.tracking.test.common.mockConnectFailure
 import com.ably.tracking.test.common.mockConnectSuccess
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
+import io.mockk.every
+import io.mockk.just
 import io.mockk.mockk
+import io.mockk.runs
 import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -25,9 +30,18 @@ class AddTrackableWorkerTest {
     private val ably: Ably = mockk {
         coEvery { startConnection() } returns Result.success(Unit)
     }
+    private val hooks: DefaultCorePublisher.Hooks = mockk {
+        every { trackables } returns null
+    }
+    private val publisherInteractor = mockk<PublisherInteractor> {
+        every { startLocationUpdates(any()) } just runs
+        every { updateTrackables(any()) } just runs
+        every { updateTrackableStateFlows(any()) } just runs
+        every { resolveResolution(any(), any()) } just runs
+    }
     private val trackable = Trackable("testtrackable")
 
-    private val worker = AddTrackableWorker(trackable, resultCallbackFunction, {}, {}, ably)
+    private val worker = AddTrackableWorker(trackable, resultCallbackFunction, {}, {}, ably, publisherInteractor, hooks)
 
     private val asyncWorks = mutableListOf<suspend () -> Unit>()
     private val postedWorks = mutableListOf<WorkerSpecification>()
@@ -50,7 +64,7 @@ class AddTrackableWorkerTest {
         assertThat(postedWorks).isEmpty()
 
         assertThat(updatedProperties.duplicateTrackableGuard.isCurrentlyAddingTrackable(trackable))
-            .isTrue()
+            .isFalse()
     }
 
     @Test
@@ -182,6 +196,173 @@ class AddTrackableWorkerTest {
             val postedWorkerSpecification = postedWorks[0] as WorkerSpecification.AddTrackableFailed
             assertThat(postedWorkerSpecification.trackable).isEqualTo(trackable)
             assertThat(postedWorkerSpecification.callbackFunction).isEqualTo(resultCallbackFunction)
+        }
+    }
+
+    @Test
+    fun `should add the trackable to the tracked trackables when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            val updatedProperties = worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            assertThat(updatedProperties.trackables).contains(trackable)
+        }
+    }
+
+    @Test
+    fun `should update the tracked trackables when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            verify(exactly = 1) {
+                publisherInteractor.updateTrackables(any())
+            }
+        }
+    }
+
+    @Test
+    fun `should calculate a resolution for the added trackable when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            verify(exactly = 1) {
+                publisherInteractor.resolveResolution(trackable, any())
+            }
+        }
+    }
+
+    @Test
+    fun `should set a state flow for the trackable when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            val updatedProperties = worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            assertThat(updatedProperties.trackableStateFlows[trackable.id]).isNotNull()
+        }
+    }
+
+    @Test
+    fun `should update state flows when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            verify(exactly = 1) {
+                publisherInteractor.updateTrackableStateFlows(any())
+            }
+        }
+    }
+
+    @Test
+    fun `should set the initial trackable state to offline when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            val updatedProperties = worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            assertThat(updatedProperties.trackableStates[trackable.id]).isInstanceOf(TrackableState.Offline::class.java)
+        }
+    }
+
+    @Test
+    fun `should call the adding trackable callback with a success when executing normally`() {
+        runTest {
+            // given
+            val initialProperties = createPublisherProperties()
+            initialProperties.duplicateTrackableGuard.clear(trackable)
+            ably.mockConnectSuccess(trackable.id)
+
+            // when
+            worker.doWork(
+                initialProperties,
+                asyncWorks.appendWork(),
+                postedWorks.appendSpecification()
+            )
+
+            // then
+            assertThat(asyncWorks).isNotEmpty()
+            assertThat(postedWorks).isEmpty()
+
+            verify { resultCallbackFunction.invoke(match { it.isSuccess }) }
         }
     }
 }
